backend/requirements.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
pillow==11.3.0
psycopg==3.2.10
psycopg-binary==3.2.10
PyJWT==2.11.0
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.5
stripe==12.2.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3